on-hold fix due to openload being unbreakable for me

+++ TrippleXMovieLoada.py
@@ -4,18 +4,25 @@ import os
 import time
 import threading
 import multiprocessing
+import tempfile
 import requests
 from PyQt5 import QtCore, QtGui
 from PyQt5.QtWidgets import *
 from PyQt5.QtCore import *
 from PyQt5.QtGui import *
+from selenium import webdriver
 
 PREVLIST = []
 NEXTLIST = []
+MAXPAGES = 1
 
 BASEURL = "http://www.freeomovie.com/page/"
 DATAPATH = str(os.getcwd() + "/Data/")
 VIDEOPATH = str(os.getcwd() + "/Video/")
+BROWSER = "firefox"
+
+HEADERS = requests.utils.default_headers()
+HEADERS.update({"User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/41.0.2227.0 Safari/537.36",})
 
 class Window():
     widget = 0
@@ -36,38 +43,39 @@ class Window():
         time.sleep(5)
         self.widget = QWidget()
         self.widget.setObjectName("Widget")
-        self.widget.setGeometry(QRect(0, 0, 756, 510))
+        self.widget.setGeometry(QRect(0, 0, 800, 600))
 
         self.curr = PREVLIST.pop(self.itera)
         PREVLIST.insert(self.itera, self.curr)
         self.note = str(self.curr[0])
         self.pxmp = QPixmap(self.curr[1])
+        self.pxmp = self.pxmp.scaledToHeight(500)
         self.url = self.curr[2]
 
         self.btn_save = QPushButton(self.widget)
         self.btn_save.setText("Save")
-        self.btn_save.setGeometry(QRect(640, 440, 104, 26))
+        self.btn_save.setGeometry(QRect(650, 440, 100, 30))
         self.btn_save.clicked.connect(lambda: self.save())
         self.btn_remove = QPushButton(self.widget)
         self.btn_remove.setText("Remove")
-        self.btn_remove.setGeometry(QRect(640, 480, 104, 26))
+        self.btn_remove.setGeometry(QRect(650, 480, 100, 30))
         self.btn_remove.clicked.connect(lambda: self.remove())
         self.btn_skip = QPushButton(self.widget)
         self.btn_skip.setText("Skip Remaining")
-        self.btn_skip.setGeometry(QRect(10, 470, 104, 26))
+        self.btn_skip.setGeometry(QRect(10, 480, 100, 30))
         self.btn_skip.clicked.connect(lambda: self.skip())
         self.btn_next = QPushButton(self.widget)
         self.btn_next.setText("Next")
-        self.btn_next.setGeometry(QRect(630, 10, 104, 26))
+        self.btn_next.setGeometry(QRect(650, 10, 100, 30))
         self.btn_next.clicked.connect(lambda: self.movelistnext())
         self.btn_prev = QPushButton(self.widget)
         self.btn_prev.setText("Previous")
-        self.btn_prev.setGeometry(QRect(10, 10, 104, 26))
+        self.btn_prev.setGeometry(QRect(10, 10, 100, 30))
         self.btn_prev.clicked.connect(lambda: self.movelistprev())
 
         self.img_label = QLabel(self.widget)
         self.img_label.setText(self.note)
-        self.img_label.setGeometry(QRect(140, 10, 471, 491))
+        self.img_label.setGeometry(QRect(100, 10, 500, 500))
         self.img_label.setPixmap(self.pxmp)
 
         QMetaObject.connectSlotsByName(self.widget)
@@ -80,6 +88,7 @@ class Window():
             self.curr = PREVLIST.pop(self.itera)
             self.note = str(self.curr[0])
             self.pxmp = QPixmap(self.curr[1])
+            self.pxmp = self.pxmp.scaledToHeight(500)
             self.url = self.curr[2]
             self.img_label.setPixmap(self.pxmp)
             print("Next!")
@@ -87,6 +96,7 @@ class Window():
             self.itera = self.itera + 1
             self.curr = PREVLIST.pop(self.itera)
             self.note = str(self.curr[0])
+            self.pxmp = self.pxmp.scaledToHeight(500)
             self.pxmp = QPixmap(self.curr[1])
             self.url = self.curr[2]
             self.img_label.setPixmap(self.pxmp)
@@ -100,6 +110,7 @@ class Window():
             self.curr = PREVLIST.pop(self.itera)
             self.note = str(self.curr[0])
             self.pxmp = QPixmap(self.curr[1])
+            self.pxmp = self.pxmp.scaledToHeight(500)
             self.url = self.curr[2]
             self.img_label.setPixmap(self.pxmp)
 
@@ -110,6 +121,7 @@ class Window():
             self.curr = PREVLIST.pop(self.itera)
             self.note = str(self.curr[0])
             self.pxmp = QPixmap(self.curr[1])
+            self.pxmp = self.pxmp.scaledToHeight(500)
             self.url = self.curr[2]
             self.img_label.setPixmap(self.pxmp)
             print("Prev!")
@@ -118,6 +130,7 @@ class Window():
             self.curr = PREVLIST.pop(self.itera)
             self.note = str(self.curr[0])
             self.pxmp = QPixmap(self.curr[1])
+            self.pxmp = self.pxmp.scaledToHeight(500)
             self.url = self.curr[2]
             self.img_label.setPixmap(self.pxmp)
             print("Prev!")
@@ -127,6 +140,7 @@ class Window():
             self.curr = PREVLIST.pop(self.itera)
             self.note = str(self.curr[0])
             self.pxmp = QPixmap(self.curr[1])
+            self.pxmp = self.pxmp.scaledToHeight(500)
             self.url = self.curr[2]
             self.img_label.setPixmap(self.pxmp)
 
@@ -141,6 +155,7 @@ class Window():
                 #self.curr = PREVLIST.pop(self.itera)
                 #self.note = str(self.curr[0])
                 #self.pxmp = QPixmap(self.curr[1])
+                #self.pxmp = self.pxmp.scaledToHeight(500)
                 #self.url = self.curr[2]
                 #self.img_label.setPixmap(self.pxmp)
                 print("finished adding")
@@ -181,7 +196,7 @@ class Window():
         if not os.path.exists(VIDEOPATH):
             os.mkdir(VIDEOPATH)
 
-        for number in range(1, 2):
+        for number in range(1, MAXPAGES+1):
             worker = threading.Thread(target=self.parser, args=(str(number), ), daemon=True)
             worker.start()
 
@@ -222,9 +237,10 @@ class Window():
 
     def my_download(self, url, outname):
         """ """
-        with open(outname, "wb") as file:#open in binary write mode
-            response = requests.get(url)#get request
-            file.write(response.content)#write to file
+        if not os.path.exists(outname):
+            with open(outname, "wb") as file:#open in binary write mode
+                response = requests.get(url, headers=HEADERS)#get request
+                file.write(response.content)#write to file
 
     def startnxtroutine(self):
         for movie in NEXTLIST:
@@ -241,7 +257,7 @@ class Window():
                 if "<li><a id=\'Openloadco\'" in line:
                     url = str(next(out))
                     url = url.split("&myURL[]=")[1].split("\"")[0]
-                    print(str(movie[0]) + str(url))
+                    print(str(movie[0]) + " " + str(url))
                     if "&myURL[]=" in url:
                         for x in range(0, (len(url.split("&myURL[]=")) - 1)):
                             partlist.append(url.split("&myURL[]=")[x])
@@ -258,7 +274,7 @@ class Window():
                 if "<li><a id=\'streamcloud\'" in line:
                     url = str(next(out))
                     url = url.split("&myURL[]=")[1].split("&tab=1\"")[0]
-                    print(str(movie[0]) + str(url))
+                    print(str(movie[0]) + " " + str(url))
                     if "&myURL[]=" in url:
                         for x in range(0, (len(url.split("&myURL[]=")) - 1)):
                             partlist.append(url.split("&myURL[]=")[x])
@@ -274,20 +290,65 @@ class Window():
 
     def parsepart(self, name, url, index):
         partname = DATAPATH + str(name) + " CD " + str(index) + ".html"
+        partfile = VIDEOPATH + str(name) + " CD " + str(index) + ".mp4"
         self.my_download(url, partname)
         if "openload" in url:
-            print("openload")
-            #use Requests to invoke button
+            url = url.replace("/f/", "/embed/")
+            subprocess.Popen([BROWSER, url, "&"]).communicate()
+            """ses = requests.Session()
+            init_req = ses.get(url, headers=HEADERS)
+            fifo = webdriver.Firefox()
+            fifo.get(url)
+            names = ["async", "http_referer", "referer"]
+            params = ["true", "", ""]
+            header2 = {"Accept" : "*/*", "Accept-Encoding" : "gzip, deflate, br", "Accept-Language" : "de,en-US;q=0.7,en;q=0.3", "Connection" : "keep-alive", "Content-Length" : "33", "Content-Type" : "application/x-www-form-urlencoded; charset=UTF-8", "DNT" : "1", "Host" : "t1.openload.co", "Origin" : "https://openload.co", "Referer" : str(url), "User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/41.0.2227.0 Safari/537.36"}
+            time.sleep(1)
+            vidurl = ""
+            vid_req = ses.post(url, data=dict(zip(names, params)), headers=header2, cookies=init_req.cookies)
+            print(vid_req.content.decode('UTF-8'))
+            with tempfile.TemporaryFile() as tempf:
+                tempf.write(vid_req.content)
+                tempf.seek(0)
+                for line in tempf:
+                    if b"<span id=\"streamurl\">" in line:
+                        print("Found URL")
+                        vidurl = line.split(b"<span id=\"streamurl\">")[1].split(b"</span>")[0]
+                        vidurl = "https://openload.co/stream/" + vidurl.decode('UTF-8')
+                        print(vidurl)
+            vidname = VIDEOPATH + str(name) + " CD " + str(index) + ".mp4"
+            print("DOwnliading")
+            self.my_download(vidurl, vidname)"""
+
+        if "streamcloud" in url:
+            ses = requests.Session()
+            init_req = ses.get(url, headers=HEADERS)
             with open(partname, "r", encoding="UTF-8") as part:
                 for line in part:
-                    if "<video id=\"" in line:
-                        vidurl = line.split("src=\"")[1].split("\">")[0]
-                        vidurl = "https://openload.co" + vidurl
-                        vidname = VIDEOPATH + str(name) + " CD " + str(index) + ".mp4"
-                        self.my_download(vidurl, vidname)
-        if "streamcloud" in url:
-            #use Requests to invoke button
-            print("StreamCloud please Fix")
+                    line = line.replace("\n", "")
+                    if "name=\"op\"" in line:
+                        op = line.split("value=\"")[1].split("\"")[0]
+                    if "name=\"usr_login\"" in line:
+                        usr_login = line.split("value=\"")[1].split("\"")[0]
+                    if "name=\"id\"" in line:
+                        _id = line.split("value=\"")[1].split("\"")[0]
+                    if "name=\"fname\"" in line:
+                        fname = line.split("value=\"")[1].split("\"")[0]
+                    if "name=\"referer\"" in line:
+                        referer = line.split("value=\"")[1].split("\"")[0]
+                    if "name=\"hash\"" in line:
+                        hash_ = line.split("value=\"")[1].split("\"")[0]
+
+            names = ["op", "usr_login", "id", "fname", "referer", "hash"]
+            params = [op, usr_login, _id, fname, referer, hash_]
+            time.sleep(12)
+            req = ses.post(url, data=dict(zip(names, params)), cookies=init_req.cookies, headers=HEADERS)
+            with tempfile.TemporaryFile() as tempf:
+                tempf.write(req.content)
+                tempf.seek(0)
+                for line in tempf:
+                    if b"file:" in line:
+                        parturl = line.split(b"file: \"")[1].split(b"\"")[0]
+                        self.my_download(parturl.decode('UTF-8'), partfile)
 
 def __main__():
